+++ src/shared/lib/stores/modal/types.ts
@@ -1,3 +1,5 @@
+import React from 'react'
+
 export interface ModalElement {
   title?: React.ReactNode
   content: React.ReactNode